+++ extensions/jetbrains/src/main/kotlin/com/deecency/udara/settings/UdaraSettings.kt
@@ -16,6 +16,7 @@ class UdaraSettings : PersistentStateComponent<UdaraSettings.State> {
         var flutterRunArgs: String = ""
         var maskSecrets: Boolean = true
         var askForDevice: Boolean = true
+        var useFlutterRunConfigs: Boolean = true
     }
 
     private var myState = State()